refactor(nuxt): within nuxt app, import directly from source file
