+++ packages/nuxt/src/app/components/layout.ts
@@ -2,7 +2,7 @@ import type { Ref, VNode } from 'vue'
 import { computed, defineComponent, h, inject, nextTick, onMounted, Transition, unref } from 'vue'
 import type { RouteLocationNormalizedLoaded } from 'vue-router'
 import { _wrapIf } from './utils'
-import { useRoute } from '#app'
+import { useRoute } from '#app/composables/router'
 // @ts-ignore
 import { useRoute as useVueRouterRoute } from '#build/pages'
 // @ts-ignore
