+++ packages/nuxt/src/app/composables/head.ts
@@ -1,0 +1,4 @@
+// eslint-disable-next-line import/no-restricted-paths
+export type { MetaObject } from '#head'
+// eslint-disable-next-line import/no-restricted-paths
+export { useHead, useSeoMeta, useServerSeoMeta } from '#head'
