+++ packages/nuxt/src/app/index.ts
@@ -7,10 +7,6 @@ export * from './config'
 
 // eslint-disable-next-line import/no-restricted-paths
 export type { PageMeta } from '../pages/runtime'
-// eslint-disable-next-line import/no-restricted-paths
-export type { MetaObject } from '../head/runtime'
-// eslint-disable-next-line import/no-restricted-paths
-export { useHead } from '#head'
 
 export const isVue2 = false
 export const isVue3 = true
