+++ packages/nuxt/src/app/plugins/cross-origin-prefetch.client.ts
@@ -1,6 +1,6 @@
 import { ref } from 'vue'
 import { parseURL } from 'ufo'
-import { defineNuxtPlugin, useHead } from '#app'
+import { defineNuxtPlugin } from '#app/nuxt'
 
 export default defineNuxtPlugin((nuxtApp) => {
   const externalURLs = ref(new Set<string>())
