+++ packages/nuxt/src/app/plugins/debug.ts
@@ -1,5 +1,5 @@
 import { createDebugger } from 'hookable'
-import { defineNuxtPlugin } from '#app'
+import { defineNuxtPlugin } from '#app/nuxt'
 
 export default defineNuxtPlugin((nuxtApp) => {
   createDebugger(nuxtApp.hooks, { tag: 'nuxt-app' })
